Use JS to Up-Filter
Javascript is now used to move between severities
whose levels are equal to or greater than the current
severity.  This way, the entire document does not have
to be retramitted for these cases.  However, the code
is written in such a way that should JS not be available
or working, normal functionality will be used.

This is commit 4/4